Add support for latest 2021.1 release
JDK11 is now required for building, so adding some notes to the readme
will help for the next build. In addition some pluin descriptions have
been updated to make the plugin installable using 2021.1.

There are some warnings coming from the lexer. That should be addressed
as well, but for now it seems to work.

Signed-off-by: Florian Bezdeka <florian@bezdeka.de>