Fix parsing of bash functions after addtask and EXPORT_FUNCTIONS
Previously the parser/highlighter failed in the following two
situations:

First:
        addtask taskname before anothertask
        bashfn() {
          :
        }

Second:
        EXPORT_FUNCTIONS functionname
        bashfn() {
          :
        }

The reason is that both lexer rules (= ADD_TASK_STATEMENT,
EXPORT_FUNC_STATEMENT) end with one BB_FUNCTION_NAME token.

The name of the bash function is also a BB_FUNCTION_NAME token and
therefor valid for the bbAddTaskStatement and bbExportFuncStatement
rules. The parser allowed the function name but failed on the opening
bracket.

Fixed by clearly terminating with CRLF tokens.

Signed-off-by: Florian Bezdeka <florian@bezdeka.de>